# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TransitOps fleet operations API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me/role:
    patch:
      operationId: setMyRole
      tags: [users]
      summary: Set the current user's role (one-time onboarding selection)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          description: Role already set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [users]
      summary: Update a user's role (Fleet Manager only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /trips:
    get:
      operationId: listTrips
      tags: [trips]
      summary: List trips
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/TripStatus"
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
    post:
      operationId: createTrip
      tags: [trips]
      summary: Create a draft trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trips/{id}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get a trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTrip
      tags: [trips]
      summary: Update a draft trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripUpdate"
      responses:
        "200":
          description: Updated trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTrip
      tags: [trips]
      summary: Delete a draft trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "400":
          description: Cannot delete a non-draft trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trips/{id}/dispatch:
    post:
      operationId: dispatchTrip
      tags: [trips]
      summary: Dispatch a draft trip (validates and assigns vehicle/driver)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dispatched trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "400":
          description: Business rule violation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trips/{id}/complete:
    post:
      operationId: completeTrip
      tags: [trips]
      summary: Complete a dispatched trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripCompleteInput"
      responses:
        "200":
          description: Completed trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "400":
          description: Business rule violation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trips/{id}/cancel:
    post:
      operationId: cancelTrip
      tags: [trips]
      summary: Cancel a dispatched trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "400":
          description: Business rule violation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    UserRole:
      type: string
      enum: [fleet_manager, driver, safety_officer, financial_analyst]
    User:
      type: object
      required: [id, clerkUserId, email, name, role, createdAt]
      properties:
        id: { type: integer }
        clerkUserId: { type: string }
        email: { type: string }
        name: { type: string }
        role:
          oneOf:
            - $ref: "#/components/schemas/UserRole"
            - type: "null"
        createdAt: { type: string, format: date-time }
    UserRoleInput:
      type: object
      required: [role]
      properties:
        role:
          $ref: "#/components/schemas/UserRole"
...
    TripStatus:
      type: string
      enum: [Draft, Dispatched, Completed, Cancelled]
    Trip:
      type: object
      required:
        [
          id,
          source,
          destination,
          vehicleId,
          driverId,
          cargoWeight,
          plannedDistance,
          status,
          actualDistance,
          fuelConsumed,
          createdAt,
          dispatchedAt,
          completedAt,
          cancelledAt,
        ]
      properties:
        id: { type: integer }
        source: { type: string }
        destination: { type: string }
        vehicleId: { type: integer }
        driverId: { type: integer }
        cargoWeight: { type: number }
        plannedDistance: { type: number }
        status:
          $ref: "#/components/schemas/TripStatus"
        actualDistance:
          type: ["number", "null"]
        fuelConsumed:
          type: ["number", "null"]
        createdAt: { type: string, format: date-time }
        dispatchedAt:
          type: ["string", "null"]
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        cancelledAt:
          type: ["string", "null"]
          format: date-time
    TripInput:
      type: object
      required:
        [source, destination, vehicleId, driverId, cargoWeight, plannedDistance]
      properties:
        source: { type: string, minLength: 1 }
        destination: { type: string, minLength: 1 }
        vehicleId: { type: integer }
        driverId: { type: integer }
        cargoWeight: { type: number, minimum: 0 }
        plannedDistance: { type: number, minimum: 0 }
    TripUpdate:
      type: object
      properties:
        source: { type: string, minLength: 1 }
        destination: { type: string, minLength: 1 }
        vehicleId: { type: integer }
        driverId: { type: integer }
        cargoWeight: { type: number, minimum: 0 }
        plannedDistance: { type: number, minimum: 0 }
    TripCompleteInput:
      type: object
      required: [finalOdometer, fuelConsumed]
      properties:
        finalOdometer: { type: number, minimum: 0 }
        fuelConsumed: { type: number, minimum: 0 }
...