# lib/api-zod/src/generated/types/userRole.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * TransitOps fleet operations API
 * OpenAPI spec version: 0.1.0
 */

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  fleet_manager: 'fleet_manager',
  driver: 'driver',
  safety_officer: 'safety_officer',
  financial_analyst: 'financial_analyst',
} as const;